LASB-956| minor refactoring and clean up
LASB-956| minor refactoring and clean up

LASB-956| revert accidental commit
