+++ crime-means-assessment/src/main/java/uk/gov/justice/laa/crime/meansassessment/dto/maatcourtdata/HardshipReviewStatus.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.meansassessment.dto.courtdata;
+package uk.gov.justice.laa.crime.meansassessment.dto.maatcourtdata;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
